% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedialandscrapeR-package.R
\docType{package}
\name{MedialandscrapeR-package}
\alias{MedialandscrapeR}
\alias{MedialandscrapeR-package}
\title{MedialandscrapeR: A simple webscraper to scrape starting pages of swiss media outlets}
\description{
A simple webscraper to scrape starting pages of swiss media outlets. The package also includes a function to do cursory analysis of newspaper-articles.
}
\keyword{internal}
