% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaplots.R
\name{mediaplots}
\alias{mediaplots}
\title{Mediaplots: Quickly analyze mediascraper-output visually}
\usage{
mediaplots(
  datacon,
  dbname = NULL,
  plots = TRUE,
  searchterm = NULL,
  searchtarget = "title"
)
}
\arguments{
\item{datacon}{Either a dataframe or a sql-connection-object that was generated by the mediascraper-function}

\item{dbname}{The name of the sql-database. Only necessary if you provided a connection-object in datacon. Default: NULL}

\item{plots}{Provides a summary of the scraped data in the form of basic plots if TRUE. Plots show: Number of articles per outlet scraped and mean-length of title/lead/body of the articles that where scraped. Default: TRUE}

\item{searchterm}{Regex-expression, string, word you want to search for. Prints a plot with the number of found appearances per media-outlet. Default: NULL}

\item{searchtarget}{Destination in which you want to search for the search term. Can either be "title", "lead" or "body". Default: "title"}
}
\value{
ggplots
}
\description{
Mediaplots: Quickly analyze mediascraper-output visually
}
\examples{
scrapingresults = data.frame(title = rep("Test", 3), lead = rep("Test", 3),
body = rep("Test", 3), outlet = c("Watson", "20 Minuten", "SRF"))
mediaplots(scrapingresults, plots = TRUE)
mediaplots(scrapingresults, plots = TRUE, searchterm = "Test")
}
