% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediascraper.R
\name{mediascraper}
\alias{mediascraper}
\title{Mediascraper: An easy way to webscrape Swiss media outlets}
\usage{
mediascraper(
  outlets,
  browser = "firefox",
  port = 4490L,
  sqldb = FALSE,
  dbname = "scrapingresults",
  plots = FALSE
)
}
\arguments{
\item{outlets}{Vector of media-outlets that should be scraped. Currently supported: Watson.ch, 20 Minuten.ch, SRF.ch, Tagesanzeiger.ch. You can also provide "all" to scrape all outlets that are currently supported}

\item{browser}{Browser that should be used for scraping. Default: Firefox}

\item{port}{Port that should be used for scraping. Default: 4491L}

\item{sqldb}{Whether scraping-results should to be stored in an sql-database or not. Default: FALSE}

\item{dbname}{Custom name for the sql-database. Default: "scrapingresults"}

\item{plots}{Provides a summary of the scraped data in the form of basic plots if TRUE. Plots show: Number of articles per outlet scraped and mean-length of title/lead/body of the articles that where scraped. Default: FALSE}
}
\value{
R-dataframe or sql-database
}
\description{
Mediascraper: An easy way to webscrape Swiss media outlets
}
\examples{
mediascraper(outlets = c("Watson", "20 Minuten"), plots = TRUE)
}
