% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medialandscraper.R
\name{mediascraper}
\alias{mediascraper}
\title{Title}
\usage{
mediascraper(
  outlets,
  browser = "firefox",
  port = 4491L,
  sqldb = FALSE,
  dbname,
  plots = FALSE,
  searchterm = NULL
)
}
\arguments{
\item{outlets}{Vector of media-outlets that should be scraped. Currently supported: Watson.ch, 20 Minuten.ch}

\item{browser}{Browser that should be used for scraping. Default: Firefox}

\item{port}{Port that should be used for scraping. Default: 4491L}

\item{sqldb}{Whether scraping-results should to be stored in an sql-database or not. Default: FALSE}

\item{dbname}{Name for the sql-database if sqldb = TRUE}

\item{plots}{Provides a summary of the scraped data in the form of basic plots if TRUE. Plots show: Number of articles per outlet scraped and mean-length of title/lead/body of the articles that where scraped. Default: FALSE}

\item{searchterm}{Regular expression for which you want to search in the titles of the scraped articles. Result is a plot}
}
\value{
R-dataframe or sql-database
}
\description{
Title
}
\examples{
mediascraper(outlets = c("Watson", "20 Minuten"), plots = TRUE)
}
